backend/requirements.txt
fastapi
uvicorn
pydantic
psycopg[binary]
Pillow



